git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@5699 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ prod/Watchdogd/Dls/The_dls.c
@@ -823,7 +823,7 @@ end:
 /* Met à jour l'entrée analogique num à partir de sa valeur avant mise a l'echelle                                            */
 /* Sortie : Néant                                                                                                             */
 /******************************************************************************************************************************/
- void Dls_data_set_AI ( gchar *tech_id, gchar *acronyme, gpointer *ai_p, float val_avant_ech, gboolean in_range )
+ void Dls_data_set_AI ( gchar *tech_id, gchar *acronyme, gpointer *ai_p, gdouble val_avant_ech, gboolean in_range )
   { struct DLS_AI *ai;
 
     if (!ai_p || !*ai_p)
